authorize user to manage their posts

+++ app/Http/Controllers/API/Post/PostController.php
@@ -8,11 +8,15 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use App\Trait\ModelNotFound;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class PostController extends Controller
 {
+    use ModelNotFound;
     /**
      * Display a listing of the resource.
      */
@@ -40,6 +44,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->modelNotFound();
 
         return new PostResource($post->load('author'));
     }
@@ -49,6 +54,9 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        $this->modelNotFound();
+
+        Gate::authorize('update', $post);
 
         $data = $request->validated();
 
@@ -62,6 +70,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->modelNotFound();
+
+        Gate::authorize('delete', $post);
+
         $post->delete();
 
         return response('', 200);
